alpine: iwd needs to be in the boot runlevel
Startup was slow because networking (in boot runlevel) was needing iwd
to start first but iwd was in the "default" runlevel which starts after
"boot". Figured this was the case because
/etc/runlevels/boot/networking has a line "provide net" in depend()
but
/etc/runlevels/default/iwd has a line "before net" in depend()
so I ran:
	rc-update delete iwd default
	rc-update add iwd boot

Still need to move these services to some kind of async runlevel so I
can get a login prompt quick before networking starts.