alpine: remove some TODO things I won't do now
I don't remember the last time I used hibernate so I'm not going to
spend time on that now.
For the changing access points problem, I'm guessing it's a bug
somewhere in the installed version of iwd. It would be fun to build and
install the latest version of iwd and if it's still a problem, debug it
and figure out what's going on but I just don't have the time right now;
in the meantime I figure I'll just disconnect and reconnect if I move
my laptop around.